Updated the encode interface to detect threshold through the salt field
The encode function was modified to allow a '$' character inside the
salt field. Such character allows for identification between threshold
and thresholdless accounts. Likewise, the verify function now uses the
same method to make the same differentiation and parse existing
hash_strings in the same way.

The following supporting functions were added:

    _polyhash_entry(password, salt, sharenumber):
        * Produces a polhashed entry
        * Is used for both, encoding and verification of passwords

    _encrypt_entry(password, salt)
        * Produces an encrypted hash
        * Is used for both, encoding and verification of passwords

Some other small errors were fixed such as the wrong sharenumber
starting value and the interface for this in the tests.