Added first sketch of the initialization command
The script added as "initialize_pph_context" now assigns the secret,
shamir secret object and thresholdless key from scratch.

The hasher assumes it is locked upon startup and will attempt to recover
this information with the verification fields or by the issuing of this
command.

The tests have been updated to include a call to the admin command to
verify it matches.

Some supporting functions have been moved from the hasher class into the
admin command implementation.