Added support for smarter recombination (using itertools.combinations)
The basic idea of this patch is to handle errors when partial
verification passes with a wrong hash in a threshold account. This error
usually appears after recombination, since the wrong hashes are not able
to provide this capability.

This patch tries to combine the secret with all the possible shares
provided. It also attempts to figure out which shares are wrong.