Fixes a wrong hash-verification routine
After recombination, we attempt to verify previous hashes. This implied
that I stored the salted-hash value of them for future verifications.
There was an inconsistency between the way I stored the salted-hash and
the way I verified it, one as a byte_array and the other as a base 64
encoded string.